+++ app/hooks/useRevalidateTag.ts
@@ -1,17 +1,17 @@
-import { revalidateTag } from "@/app/[company]/actions";
+import { revalidateTagAction } from "@/app/[company]/actions";
 import { getRecaptchaToken, Action } from "@/app/lib/getRecapchaToken";
 import { useTransition } from "react";
 
 export function useRevalidateTag({ tag }: { tag: string }) {
   const [isPending, startTransition] = useTransition();
-  const onClickHandler = async () => {
+  const revalidateTag = async () => {
     startTransition(async () => {
       const recaptchaToken = await getRecaptchaToken(Action.REVALIDATE_TAG);
-      await revalidateTag({
+      await revalidateTagAction({
         tag,
         recaptchaToken,
       });
     });
   };
-  return [isPending, onClickHandler] as const;
+  return [isPending, revalidateTag] as const;
 }
